feat(utils): Add global bypass functionality
This commit adds functionality to bypass certain requirements in the Giveaway class. A new condition is introduced that checks if the user has any role specified in the global bypass list. If so, the user is considered to have bypassed the requirements. Additionally, a new button labeled "Global Bypass" is added to the GiveawayConfig class, allowing users to add or remove roles from the global bypass list.